Struct egui_extras::image::RetainedImage

pub struct RetainedImage { /* private fields */ }

An image to be shown in egui.

Load once, and save somewhere in your app state.

Use the svg and image features to enable more constructors.

Implementations

impl RetainedImage

pub fn from_color_image(debug_name: impl Into<String>, image: ColorImage) -> Self

pub fn from_image_bytes(
debug_name: impl Into<String>,
image_bytes: &[u8]
) -> Result<Self, String>

Load a (non-svg) image.

image_bytes should be the raw contents of an image file (.png, .jpg, โฆ).

Requires the โimageโ feature. You must also opt-in to the image formats you need with e.g. image = { version = "0.24", features = ["jpeg", "png"] }.

Errors

On invalid image or unsupported image format.

pub fn with_options(self, options: TextureOptions) -> Self

Set the texture filters to use for the image.

Note: If the texture has already been uploaded to the GPU, this will require re-uploading the texture with the updated filter.

Example

// Upload a pixel art image without it getting blurry when resized
let image = RetainedImage::from_color_image("my_image", color_image)
    .with_options(TextureOptions::NEAREST);

pub fn with_texture_filter(self, filter: TextureFilter) -> Self

๐Deprecated: Use with_options instead

pub fn size(&self) -> [usize; 2]

The size of the image data (number of pixels wide/high).

pub fn width(&self) -> usize

The width of the image.

pub fn height(&self) -> usize

The height of the image.

pub fn size_vec2(&self) -> Vec2

The size of the image data (number of pixels wide/high).

pub fn debug_name(&self) -> &str

The debug name of the image, e.g. the file name.

pub fn texture_id(&self, ctx: &Context) -> TextureId

The texture if for this image.

pub fn show_max_size(&self, ui: &mut Ui, max_size: Vec2) -> Response

Show the image with the given maximum size.

pub fn show(&self, ui: &mut Ui) -> Response

Show the image with the original size (one image pixel = one gui point).

pub fn show_scaled(&self, ui: &mut Ui, scale: f32) -> Response

Show the image with the given scale factor (1.0 = original size).

pub fn show_size(&self, ui: &mut Ui, desired_size: Vec2) -> Response

Show the image with the given size.