egui_extras

Struct StripBuilder

pub struct StripBuilder<'a> { /* private fields */ }

Builder for creating a new Strip.

This can be used to do dynamic layouts.

In contrast to normal egui behavior, strip cells do not grow with its children!

First use Self::size and Self::sizes to allocate space for the rows or columns will follow. Then build the strip with Self::horizontal/Self::vertical, and add ‘cells’ to it using Strip::cell. The number of cells MUST match the number of pre-allocated sizes.

Example

use egui_extras::{StripBuilder, Size};
StripBuilder::new(ui)
    .size(Size::remainder().at_least(100.0)) // top cell
    .size(Size::exact(40.0)) // bottom cell
    .vertical(|mut strip| {
        // Add the top 'cell'
        strip.cell(|ui| {
            ui.label("Fixed");
        });
        // We add a nested strip in the bottom cell:
        strip.strip(|builder| {
            builder.sizes(Size::remainder(), 2).horizontal(|mut strip| {
                strip.cell(|ui| {
                    ui.label("Top Left");
                });
                strip.cell(|ui| {
                    ui.label("Top Right");
                });
            });
        });
    });

Implementations

impl<'a> StripBuilder<'a>

pub fn new(ui: &'a mut Ui) -> Self

Create new strip builder.

pub fn clip(self, clip: bool) -> Self

Should we clip the contents of each cell? Default: false.

pub fn cell_layout(self, cell_layout: Layout) -> Self

What layout should we use for the individual cells?

pub fn sense(self, sense: Sense) -> Self

What should strip cells sense for? Default: egui::Sense::hover().

pub fn size(self, size: Size) -> Self

Allocate space for one column/row.

pub fn sizes(self, size: Size, count: usize) -> Self

Allocate space for several columns/rows at once.

pub fn horizontal<F>(self, strip: F) -> Response
where F: for<'b> FnOnce(Strip<'a, 'b>),

Build horizontal strip: Cells are positions from left to right. Takes the available horizontal width, so there can’t be anything right of the strip or the container will grow slowly!

Returns a egui::Response for hover events.

pub fn vertical<F>(self, strip: F) -> Response
where F: for<'b> FnOnce(Strip<'a, 'b>),

Build vertical strip: Cells are positions from top to bottom. Takes the full available vertical height, so there can’t be anything below of the strip or the container will grow slowly!

Returns a egui::Response for hover events.