Struct egui_extras::TableRow

pub struct TableRow<'a, 'b> { /* private fields */ }

The row of a table. Is created by TableRow for each created TableBody::row or each visible row in rows created by calling TableBody::rows.

Implementations

impl<'a, 'b> TableRow<'a, 'b>

pub fn col( &mut self, add_cell_contents: impl FnOnce(&mut Ui), ) -> (Rect, Response)

Add the contents of a column on this row (i.e. a cell).

Returns the used space (min_rect) plus the Response of the whole cell.

pub fn set_selected(&mut self, selected: bool)

Set the selection highlight state for cells added after a call to this function.

pub fn set_hovered(&mut self, hovered: bool)

Set the hovered highlight state for cells added after a call to this function.

pub fn response(&self) -> Response

Returns a union of the Responses of the cells added to the row up to this point.

You need to add at least one row to the table before calling this function.

pub fn index(&self) -> usize

Returns the index of the row.

pub fn col_index(&self) -> usize

Returns the index of the column. Incremented after a column is added.