Crate egui_extras

This is a crate that adds some features on top top of egui.

This crate are for experimental features, and features that require big dependencies that does not belong in egui.

Feature flags

Modules

syntax_highlighting

Syntax highlighting for code.

Structs

Column

Specifies the properties of a column, like its width range.

Strip

A Strip of cells which go in one direction. Each cell has a fixed size. In contrast to normal egui behavior, strip cells do not grow with its children!

StripBuilder

Builder for creating a new Strip.

Table

Table struct which can construct a TableBody.

TableBody

The body of a table.

TableBuilder

Builder for a Table with (optional) fixed header and scrolling body.

TableRow

The row of a table. Is created by TableRow for each created TableBody::row or each visible row in rows created by calling TableBody::rows.

Enums

Size

Size hint for table column/strip cell.

Functions

install_image_loaders

Installs a set of image loaders.